requirements.txt
PyGithub>=2.1.1
streamlit>=1.28.0
python-telegram-bot>=20.6
APScheduler>=3.10.4
openai>=1.3.0
requests>=2.31.0
aiosqlite>=0.19.0
python-dateutil>=2.8.2
